ARM: dts: sun6i: The blue led on the Mele I7 is a power led
The blue led on the Mele I7 is wired to light up as soon as the board
has powered (it will be on while the gpio is still in input / floating
mode), also its location on the top-set box clearly signals "power led".

Until now we've been treating this as a generic usr function led, which
means that when you plug power into the top-set box, the power-led lights
and then turns off as soon as the kernel loads, which looks wrong.

This renames the led from i7:blue:usr to i7:blue:pwr and marks
it as default on, fixing this.

Signed-off-by: Marcus Cooper <codekipper@gmail.com>