ARM: dts: sun8i: Extend Orange Pi Plus dt to also handle the Plus 2E
There is a new Orange Pi Plus 2E available now, rather then adding yet
another dts file for this variant, extend the existing Plus / Plus2 dt
to also support this variant.

The 2E goes from 4 to 3 USB ports and drops the sata connector, so it
no longer has a USB hub or an USB <-> sata bridge on the board.

This commit enables all 3 ehci and ohci controllers to support the 3
USB connectors which are now directly connected to the SoC.

Downside of this approach is 4 unused USB busses now showing up on
the original Plus / Plus 2.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>