ARM: dts: sun4i: Change ba10 tvbox usb_otg dr_mode to otg
Although the ba10 tvbox hs the musb connected to an USB-A connector,
it is fully otg capable including Vbus detect and control, it just lacks
an id-pin.

Add support for Vbus detection so that we will not turn on Vbus if
an USB A <-> cable is plugged in. Also put the controller in otg
mode so that it can be switched to peripheral mode using software
mode switching through the musb controllers "mode" sysfs attribute.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>