ASoC: sun8i-codec: Remove incorrect SND_SOC_DAIFMT_DSP_B
DSP_A and DSP_B are not interchangeable. The timing used by the codec in
DSP mode is consistent with DSP_A. This is verified with an EG25-G modem
connected to AIF2, as well as by comparing with the BSP driver.

Remove the DSP_B option, as it is not supported by the hardware.

Cc: stable@kernel.org
Fixes: 36c684936fae ("ASoC: Add sun8i digital audio codec")
Signed-off-by: Samuel Holland <samuel@sholland.org>