media: cedrus: Fix H265 aux buffer size
Neighbour info buffer size needs to be 794 kiB in size on H6. This is
actually mentioned in comment, but smaller size is used nevertheless.

Increase buffer size to conform H6 needs. Since increase is not that big
in absolute numbers, it doesn't make sense to complicate logic for older
generations.

Bug was discovered using iommu, which reported access error when trying
to play H265 video.

Fixes: 86caab29da78 ("media: cedrus: Add HEVC/H.265 decoding support")
Signed-off-by: Jernej Skrabec <jernej.skrabec@gmail.com>